Patch Jbuilder to Support Varying Cache
Previously, admins were not able to see up-to-date data in API requests
due to the `#cache!` method in Jbuilder not being patched to skip
caching when an admin is logged in. To resolve this, Workarea now
applies the same patch to Jbuilder as it does to ActionView. Reading
from the cache is now skipped if you're logged in as an admin, and cache
keys are appended with the configured `Cache::Varies` just the same as
in regular Haml views.

WORKAREA-243